requirements.txt
pydoll-python>=0.1.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
pytest>=7.0.0
